Add Newton Third implementation.
Newton's Third Law means we can halve the amount of computation at the cost of
increasing the amount of communication. We use an inverse map (from particles to
elements within the neighbor list) and a gather kernel. Other implementation
options are using atomics.

The neighbor list can be made into four implementations:
{TpA, BpA} x {NoNewton, WithNewton}